'Observer'

+++ app/Http/Controllers/Admin/Langs/LangControll.php
@@ -17,24 +17,30 @@ class LangControll extends Controller
         $admin=Auth::guard('admin')->user();
         return view('admin.langs.add_lang',compact('admin'));
     }
-    public function store(ValidLang $req){
+    public function store(ValidLang $request){
 
        // return response()->json($req->all());
-        $statue=$req->statue;
-        $data=$req->except('_token');
+
         try{
-            Lang::create([
-                'abbr'=>$data['abbr'],
-                'name'=>$data['name'],
-                'direction'=>$data['direction'],
-                'action'=>$statue[0]
-            ]);
+            if(isset($request) && !empty($request)) {
+                $statue = $request->statue;
+                $data = $request->except('_token');
+                Lang::create([
+                    'abbr' => $data['abbr'],
+                    'name' => $data['name'],
+                    'direction' => $data['direction'],
+                    'statue' => $statue[0]
+                ]);
+                return response()->json([
+                    'statue' => true,
+                    'msg' => 'created done'
+                ]);
+            }
             return response()->json([
-                'statue'=>true,
-                'msg'=>'created done'
+                'statue'=>false,
+                'msg'=>'Request not  found'
             ]);
         }catch(\Exception $ex){
-            //return $ex;
             return response()->json([
                 'statue'=>false,
                 'msg'=>'created falis'
@@ -42,29 +48,27 @@ class LangControll extends Controller
         }
     }
     public function available_langs(){
-        $langs= Lang::Data();
+        $langs= Lang::Data()->paginate(paginate_count);
         $admin=Auth::guard('admin')->user();
         $user=Auth::guard('web')->user();
         return view('admin.langs.available_langs',compact('langs','admin','user'));
     }
 
-       public function delete($lang_id){
-           $data=Lang::where('id',$request->id)->first();
-           if(isset($data) && $data!=null){
-               $data->delete();
-               return response()->json([
-                   'statue'=>true,
-                   'msg'=>'Deleted Done'
-               ]);
-           }
-           return response()->json([
-               'statue'=>false,
-               'msg'=>'Deleted fails'
-           ]);
-
-//        $lang=Lang::find($lang_id);
-//        $lang_id->delete();
-//        return redirect(route('available_langs'))->with('success','Deleted Done');
+       public function delete($request){
+        if(isset($request) && !empty($request)) {
+            $data = Lang::where('id', $request->id)->first();
+            if (isset($data) && $data != null) {
+                $data->delete();
+                return response()->json([
+                    'statue' => true,
+                    'msg' => 'Deleted Done'
+                ]);
+            }
+            return response()->json([
+                'statue'=>false,
+                'msg'=>'Deleted fails'
+            ]);
+        }
 
     }
 
@@ -92,19 +96,14 @@ class LangControll extends Controller
 
     }
 
-//    public function delete_ajax(Request $request){
-//        $data=Lang::where('id',$request->id)->first();
-//        if(isset($data) && $data!=null){
-//            $data->delete();
-//            return response()->json([
-//                'statue'=>true,
-//                'msg'=>'Deleted Done'
-//            ]);
-//        }
-//        return response()->json([
-//         'statue'=>false,
-//         'msg'=>'Deleted fails'
-//     ]);
-//     }
+        public function change_statue($lang_id){
+                $lang= Lang::find($lang_id);
+                if(isset($lang) && $lang !=null){
+                   $statue=$lang->statue==1 ? 0 : 1 ;
+                    $lang->update(['statue'=>$statue]);
+                    return redirect(route('available_langs'))->with('success','Langs ia '. $lang->getStatue());
+                }
+            return redirect(route('avalibale_langs'))->with('error','Not Found');
+        }
 }
